+++ SecureVault/src/controllers/internalShare.controller.js
@@ -61,19 +61,30 @@ class InternalShareController {
       if (result.iv) res.setHeader("X-File-IV", result.iv);
     }
 
+    if (result.isMasked) {
+      res.setHeader("X-Is-Masked", "true");
+    }
+
+    let cleanedUp = false;
+    const cleanup = () => {
+      if (!cleanedUp && result.path) {
+        cleanedUp = true;
+        fs.unlink(result.path, () => {});
+      }
+    };
+
     const fileStream = fs.createReadStream(result.path);
     fileStream.pipe(res);
 
     fileStream.on("error", (err) => {
-      fs.unlink(result.path, () => {});
+      cleanup();
       if (!res.headersSent) {
         res.status(500).json({ success: false, message: "Error streaming file" });
       }
     });
 
-    res.on("finish", () => {
-      fs.unlink(result.path, () => {});
-    });
+    res.on("finish", cleanup);
+    res.on("close", cleanup);
   });
 }
